requirements.txt
datasets>=2.18.0
transformers>=4.40.0
sentence-transformers>=2.6.0
faiss-cpu>=1.7.4
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
jsonlines>=3.1.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
